refactor: Split binary persistence into logical submodules
- Create binary/ directory structure with five focused modules
- format.rs: File format constants, version, and TypeTag enum
- io.rs: Low-level read/write primitives for all types
- value.rs: SqlValue serialization/deserialization
- catalog.rs: Schema, table, index, and role serialization
- data.rs: Table row data serialization
- mod.rs: Public API and compression handling

Benefits:
- Better separation of concerns (each module has single responsibility)
- Improved readability (1000 line file split into 6 focused modules)
- Easier maintenance (format changes isolated to relevant modules)
- Better testability (each module can be tested independently)
- No functional changes or API breakage

All existing tests pass. Lines reduced from 934 to 1024 total
(includes documentation and test code in each module).
